EBOOK_PROJECT/requirements.txt
asgiref==3.5.2
Django==3.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
Markdown==3.4.1
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.3
